+++ src/main/java/com/skrra/atmosphereplus/client/AtmospherePlusClient.java
@@ -3,6 +3,7 @@ package com.skrra.atmosphereplus.client;
 import com.skrra.atmosphereplus.automation.BiomeAtmosphereManager;
 import com.skrra.atmosphereplus.config.ConfigManager;
 import com.skrra.atmosphereplus.config.ConfigSafety;
+import com.skrra.atmosphereplus.environment.EnvironmentDetector;
 import com.skrra.atmosphereplus.keybind.AtmosphereKeybinds;
 import com.skrra.atmosphereplus.presets.PresetLibraryManager;
 import com.skrra.atmosphereplus.themes.ThemeManager;
@@ -12,7 +13,7 @@ import net.fabricmc.api.ClientModInitializer;
 public class AtmospherePlusClient implements ClientModInitializer {
     public static final String MOD_ID = "atmosphereplus";
     public static final String MOD_NAME = "Atmosphere+";
-    public static final String VERSION = "0.3.0-beta.4";
+    public static final String VERSION = "0.3.0-beta.5";
 
     @Override
     public void onInitializeClient() {
@@ -21,6 +22,7 @@ public class AtmospherePlusClient implements ClientModInitializer {
         PresetLibraryManager.init();
         ThemeManager.init();
         AtmosphereKeybinds.register();
+        EnvironmentDetector.register();
         TransitionManager.register();
         BiomeAtmosphereManager.register();
     }